requirements.txt
fastapi==0.115.8
grpcio==1.67.1
httpx==0.28.1
protobuf==5.29.3
pydantic==2.10.6
python-dotenv==1.0.1
